backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# AI/LLM Integration
google-genai==0.3.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
openpyxl==3.1.2
PyPDF2==3.0.1

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
cryptography==42.0.0
pyjwt==2.8.0

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0

# Rate Limiting
slowapi==0.1.9
